+++ src/FreeElephants/RestDaemon/Endpoint/EndpointMethodHandlerInterface.php
@@ -3,6 +3,7 @@
 namespace FreeElephants\RestDaemon\Endpoint;
 
 use FreeElephants\RestDaemon\Middleware\Collection\EndpointMiddlewareCollectionInterface;
+use FreeElephants\RestDaemon\Middleware\MiddlewareInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UriInterface;
@@ -10,9 +11,14 @@ use Psr\Http\Message\UriInterface;
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
-interface EndpointMethodHandlerInterface
+interface EndpointMethodHandlerInterface extends MiddlewareInterface
 {
 
+    /**
+     * @deprecated will be make protected or private.
+     * @param ServerRequestInterface $request
+     * @return ResponseInterface
+     */
     public function handle(ServerRequestInterface $request): ResponseInterface;
 
     public function setMiddlewareCollection(EndpointMiddlewareCollectionInterface $endpointMiddlewareCollection);